feature(outdoor) : 실외지도 수정 및 초기 위치 수정
