+++ apps/core/src/app/store/facilityStore.ts
@@ -1,12 +1,10 @@
 import { create } from 'zustand'
 import { devtools } from 'zustand/middleware'
-import { facilityService, domainUtils } from '@plug/common-services'
+import { facilityService } from '@plug/common-services'
 import type { FacilityResponse, FacilityType } from '@plug/common-services'
 
-type FacilitiesData = Record<string, FacilityResponse[]>
-
 interface FacilityState {
-  facilities: FacilitiesData
+  facilities: FacilityResponse[]
   error: string | null
   selectedFacility: FacilityResponse | null
   facilitiesFetched: boolean
@@ -14,7 +12,7 @@ interface FacilityState {
 }
 
 interface FacilityActions {
-  setFacilities: (facilities: FacilitiesData) => void
+  setFacilities: (facilities: FacilityResponse[]) => void
   setError: (error: string | null) => void
   setSelectedFacility: (facility: FacilityResponse | null) => void
   setFacilitiesFetched: (fetched: boolean) => void
@@ -31,38 +29,20 @@ let loadFacilitiesPromise: Promise<void> | null = null
 export const useFacilityStore = create<FacilityStore>()(
   devtools(
     (set, get) => ({
-      facilities: {},
+      facilities: [],
       error: null,
       selectedFacility: null,
-  facilitiesFetched: false,
-  isLoading: false,
+      facilitiesFetched: false,
+      isLoading: false,
 
       setFacilities: (facilities) => set({ facilities }),
       setError: (error) => set({ error }),
       setSelectedFacility: (selectedFacility) => set({ selectedFacility }),
       setFacilitiesFetched: (facilitiesFetched) => set({ facilitiesFetched }),
 
-      getAllFacilities: () => {
-        const { facilities } = get()
-        return Object.values(facilities)
-          .filter(Array.isArray)
-          .flat() as FacilityResponse[]
-      },
-
-      getFacilityById: (id) => {
-        const { facilities } = get()
-        const allFacilities = Object.values(facilities)
-          .filter(Array.isArray)
-          .flat() as FacilityResponse[]
-        return allFacilities.find(facility => facility.id === id)
-      },
-
-      getFacilitiesByType: (type: FacilityType) => {
-        const { facilities } = get()
-        const endpointKey = domainUtils.getConfig(type).endpoint
-        const list = facilities[endpointKey]
-        return Array.isArray(list) ? list : []
-      },
+      getAllFacilities: () => get().facilities,
+      getFacilityById: (id) => get().facilities.find(f => f.id === id),
+      getFacilitiesByType: (type: FacilityType) => get().facilities.filter(f => (f as FacilityResponse & { type?: FacilityType }).type === type),
 
       loadFacilities: async () => {
         const { facilitiesFetched, isLoading } = get()
@@ -74,18 +54,8 @@ export const useFacilityStore = create<FacilityStore>()(
         loadFacilitiesPromise = (async () => {
           try {
             const response = await facilityService.getAllFacilities()
-
-            const flatList = Array.isArray(response.data) ? response.data : []
-            const groupedByEndpoint: FacilitiesData = {}
-
-            flatList.forEach((item) => {
-              const itemType = (item as FacilityResponse & { type?: FacilityType }).type
-              const endpoint = itemType ? domainUtils.getConfig(itemType).endpoint : domainUtils.getConfig('BUILDING').endpoint
-              if (!groupedByEndpoint[endpoint]) groupedByEndpoint[endpoint] = []
-              groupedByEndpoint[endpoint].push(item)
-            })
-
-            set({ facilities: groupedByEndpoint, facilitiesFetched: true })
+            const list = Array.isArray(response.data) ? (response.data as FacilityResponse[]) : []
+            set({ facilities: list, facilitiesFetched: true })
           } catch (error) {
             set({ error: error instanceof Error ? error.message : 'Failed to load facilities' })
           } finally {
@@ -97,9 +67,7 @@ export const useFacilityStore = create<FacilityStore>()(
         return loadFacilitiesPromise
       }
     }),
-    {
-      name: 'facility-store',
-    }
+    { name: 'facility-store' }
   )
 )
