+++ apps/core/src/app/view/components/index.ts
@@ -1,4 +1,3 @@
 export { default as MainMap } from './maps/MainMap'
 export { default as OutdoorMap } from './maps/OutdoorMap'
 export { default as IndoorMap } from './maps/IndoorMap'
-export { default as CctvModal } from './CctvModal'
