+++ packages/common-services/src/constants/domain-config.ts
@@ -1,36 +1,3 @@
-import { 
-  FacilityCreateRequest,
-  FacilityUpdateRequest, 
-  FacilityResponse,
-  FloorRequest,
-  FloorResponse,
-  StationInfoResponse,
-  StationInfoRequest,
-  StationInfoCreateRequest
-} from '../types/facility';
-
-export interface DataComponentTypes {
-  facility: {
-    createRequest: { facility: FacilityCreateRequest; };
-    updateRequest: { facility: FacilityUpdateRequest; };
-    response: { facility: FacilityResponse; };
-  };
-  floors: {
-    createRequest: { floors?: FloorRequest[]; };
-    updateRequest: { floors: FloorRequest[]; };
-    response: { floors: FloorResponse[]; };
-  };
-  stationInfo: {
-    createRequest: { stationInfo?: StationInfoCreateRequest; };
-    updateRequest: { stationInfo: StationInfoRequest; };
-    response: { stationInfo: StationInfoResponse; };
-  };
-  boundary: {
-    createRequest: { boundary?: string; };
-    updateRequest: { boundary: string; };
-    response: { boundary: string; };
-  };
-}
 
 export const DOMAINS = {
   building: {
@@ -53,45 +20,10 @@ export const DOMAINS = {
 
 } as const;
 
-type ComponentName = keyof DataComponentTypes;
-
-type ComposeCreateRequest<T extends readonly ComponentName[]> = 
-  T extends readonly [infer First, ...infer Rest]
-    ? First extends ComponentName
-      ? Rest extends readonly ComponentName[]
-        ? DataComponentTypes[First]['createRequest'] & ComposeCreateRequest<Rest>
-        : DataComponentTypes[First]['createRequest']
-      : {}
-    : {};
-
-type ComposeUpdateRequest<T extends readonly ComponentName[]> = 
-  T extends readonly [infer First, ...infer Rest]
-    ? First extends ComponentName
-      ? Rest extends readonly ComponentName[]
-        ? DataComponentTypes[First]['updateRequest'] & ComposeUpdateRequest<Rest>
-        : DataComponentTypes[First]['updateRequest']
-      : {}
-    : {};
-
-type ComposeResponse<T extends readonly ComponentName[]> = 
-  T extends readonly [infer First, ...infer Rest]
-    ? First extends ComponentName
-      ? Rest extends readonly ComponentName[]
-        ? DataComponentTypes[First]['response'] & ComposeResponse<Rest>
-        : DataComponentTypes[First]['response']
-      : {}
-    : {};
+type ComponentName = 'facility' | 'floors' | 'stationInfo' | 'boundary'
 
 export type FacilityKey = keyof typeof DOMAINS;
 export type FacilityType = Uppercase<FacilityKey>;
-
-type FacilityMap = {
-  [K in FacilityKey as Uppercase<K>]: typeof DOMAINS[K]
-};
-
-export type DomainCreateRequest<T extends FacilityType> = ComposeCreateRequest<FacilityMap[T]['components']>;
-export type DomainUpdateRequest<T extends FacilityType> = ComposeUpdateRequest<FacilityMap[T]['components']>;
-export type DomainResponse<T extends FacilityType> = ComposeResponse<FacilityMap[T]['components']>;
 
 export const domainUtils = {
   getConfig(domain: FacilityType | FacilityKey) {
@@ -108,7 +40,7 @@ export const domainUtils = {
   },
 
   hasComponent(domain: FacilityType | FacilityKey, componentName: ComponentName): boolean {
-  const config = this.getConfig(domain as FacilityType);
-  return (config.components as readonly ComponentName[]).includes(componentName);
+    const config = this.getConfig(domain as FacilityType);
+    return (config.components as readonly ComponentName[]).includes(componentName);
   }
 };
